Finished implementing Player portion of API.
PlayerInstances now change when a player changes WorldGroup.
PlayerChangeGroupEvent is now fired when a player changes WorldGroup.
PlayerInstances are saved and loaded when a player joins and leaves.
WorldGroups now can specify defaults for a PlayerInstance if an instance is not found for the group.
Added default PlayerData.yml to be used it a WorldGroup does not contain defaults.